backend/backend/requirements.txt

fastapi
uvicorn
langchain
faiss-cpu
openai
sentence-transformers
python-dotenv
